Make _bolt_optimize_file non fatal in the llvm ebuilds

+++ sys-devel/llvm/llvm-14.0.0.9999.ebuild
@@ -1249,7 +1249,7 @@ _bolt_optimize_file() {
 		llvm-bolt ${args[@]} || true
 	fi
 
-	mv "${f}{.bolt,}" || die
+	[[ -e "${f}.bolt" ]] && mv "${f}{.bolt,}" || die
 }
 
 strip_package() {
